Fix for Magento 1.8 name clash
The name of "index" for the layout update file definition, clashes with Mage_Index module. This leads to the Admin > System > Index Management page being blank.

This is due to Magento switching to using xml layout definition, from creating the block on the fly in the controller, within the Mage Index module.